server/requirements.txt
Flask
flask-cors
pymongo
python-dotenv
gunicorn
requests
google-generativeai
cloudinary